+++ core/src/main/kotlin/net/corda/core/contracts/FungibleAsset.kt
@@ -1,7 +1,12 @@
 package net.corda.core.contracts
 
+import net.corda.core.crypto.Party
 import net.corda.core.flows.FlowException
+import net.corda.core.transactions.TransactionBuilder
+import net.corda.core.utilities.loggerFor
+import net.corda.core.utilities.trace
 import java.security.PublicKey
+import java.util.*
 
 class InsufficientBalanceException(val amountMissing: Amount<*>) : FlowException("Insufficient balance, missing $amountMissing")
