+++ finance/src/main/kotlin/net/corda/contracts/clause/AbstractConserveAmount.kt
@@ -1,12 +1,11 @@
 package net.corda.contracts.clause
 
+import net.corda.contracts.asset.OnLedgerAsset
 import net.corda.core.contracts.*
 import net.corda.core.contracts.clauses.Clause
 import net.corda.core.transactions.TransactionBuilder
-import java.security.PublicKey
 import net.corda.core.utilities.loggerFor
-import net.corda.core.utilities.trace
-import java.util.*
+import java.security.PublicKey
 
 /**
  * Standardised clause for checking input/output balances of fungible assets. Requires that a
@@ -20,34 +19,6 @@ abstract class AbstractConserveAmount<S : FungibleAsset<T>, C : CommandData, T :
     }
 
     /**
-     * Gather assets from the given list of states, sufficient to match or exceed the given amount.
-     *
-     * @param acceptableCoins list of states to use as inputs.
-     * @param amount the amount to gather states up to.
-     * @throws InsufficientBalanceException if there isn't enough value in the states to cover the requested amount.
-     */
-    @Throws(InsufficientBalanceException::class)
-    private fun gatherCoins(acceptableCoins: Collection<StateAndRef<S>>,
-                            amount: Amount<T>): Pair<ArrayList<StateAndRef<S>>, Amount<T>> {
-        require(amount.quantity > 0) { "Cannot gather zero coins" }
-        val gathered = arrayListOf<StateAndRef<S>>()
-        var gatheredAmount = Amount(0, amount.token)
-        for (c in acceptableCoins) {
-            if (gatheredAmount >= amount) break
-            gathered.add(c)
-            gatheredAmount += Amount(c.state.data.amount.quantity, amount.token)
-        }
-
-        if (gatheredAmount < amount) {
-            log.trace { "Insufficient balance: requested $amount, available $gatheredAmount" }
-            throw InsufficientBalanceException(amount - gatheredAmount)
-        }
-
-        log.trace { "Gathered coins: requested $amount, available $gatheredAmount, change: ${gatheredAmount - amount}" }
-        return Pair(gathered, gatheredAmount)
-    }
-
-    /**
      * Generate an transaction exiting fungible assets from the ledger.
      *
      * @param tx transaction builder to add states and commands to.
@@ -56,41 +27,14 @@ abstract class AbstractConserveAmount<S : FungibleAsset<T>, C : CommandData, T :
      * the responsibility of the caller to check that they do not attempt to exit funds held by others.
      * @return the public key of the assets issuer, who must sign the transaction for it to be valid.
      */
+    @Deprecated("This function will be removed in a future milestone", ReplaceWith("OnLedgerAsset.generateExit()"))
     @Throws(InsufficientBalanceException::class)
     fun generateExit(tx: TransactionBuilder, amountIssued: Amount<Issued<T>>,
                      assetStates: List<StateAndRef<S>>,
                      deriveState: (TransactionState<S>, Amount<Issued<T>>, PublicKey) -> TransactionState<S>,
                      generateMoveCommand: () -> CommandData,
-                     generateExitCommand: (Amount<Issued<T>>) -> CommandData): PublicKey {
-        val owner = assetStates.map { it.state.data.owner }.toSet().singleOrNull() ?: throw InsufficientBalanceException(amountIssued)
-        val currency = amountIssued.token.product
-        val amount = Amount(amountIssued.quantity, currency)
-        var acceptableCoins = assetStates.filter { ref -> ref.state.data.amount.token == amountIssued.token }
-        tx.notary = acceptableCoins.firstOrNull()?.state?.notary
-        // TODO: We should be prepared to produce multiple transactions exiting inputs from
-        // different notaries, or at least group states by notary and take the set with the
-        // highest total value
-        acceptableCoins = acceptableCoins.filter { it.state.notary == tx.notary }
-
-        val (gathered, gatheredAmount) = gatherCoins(acceptableCoins, amount)
-        val takeChangeFrom = gathered.lastOrNull()
-        val change = if (takeChangeFrom != null && gatheredAmount > amount) {
-            Amount(gatheredAmount.quantity - amount.quantity, takeChangeFrom.state.data.amount.token)
-        } else {
-            null
-        }
-
-        val outputs = if (change != null) {
-            // Add a change output and adjust the last output downwards.
-            listOf(deriveState(gathered.last().state, change, owner))
-        } else emptyList()
-
-        for (state in gathered) tx.addInputState(state)
-        for (state in outputs) tx.addOutputState(state)
-        tx.addCommand(generateMoveCommand(), gathered.map { it.state.data.owner })
-        tx.addCommand(generateExitCommand(amountIssued), gathered.flatMap { it.state.data.exitKeys })
-        return amountIssued.token.issuer.party.owningKey
-    }
+                     generateExitCommand: (Amount<Issued<T>>) -> CommandData): PublicKey
+    = OnLedgerAsset.generateExit(tx, amountIssued, assetStates, deriveState, generateMoveCommand, generateExitCommand)
 
     override fun verify(tx: TransactionForContract,
                         inputs: List<S>,
